splitting text in subdirectory /priv/text
